tilt: rename/organize codegen makefile targets

+++ scripts/update-codegen.sh
@@ -5,6 +5,14 @@ set -e
 DIR=$(dirname "$0")
 cd "$DIR/.."
 
+
+# docker mounts don't work in our CI setup - just run the scripts directly
+if [[ $CI == true ]]; then
+  scripts/update-protobuf-helper.sh
+  scripts/update-codegen-helper.sh
+  exit 0
+fi
+
 docker build -t tilt-protobuf-helper -f scripts/protobuf-helper.dockerfile scripts
 docker run --rm -v "$(pwd)":/go/src/github.com/tilt-dev/tilt \
    --entrypoint /go/src/github.com/tilt-dev/tilt/scripts/update-protobuf-helper.sh \
